+++ language/german_utf-8.php
@@ -459,7 +459,8 @@ $LANG_confignames['evlist'] = array(
     'event_passing'         => 'Nicht mehr "Bevorstehend"',
     'default_permissions'   => '(Eigentümer,Gruppe,Benutzer,Gast)',
     'reminder_days'         => 'Anzahl der Tage vor einem Ereignis, um Erinnerungen zu erlauben',
-    'displayblocks'         => 'Blöcke anzeigen',
+    'displayblocks'         => 'Display Blocks in List Views',
+    'displayblocks_blk'     => 'Display Blocks in Month/Year Views',
     'default_view'          => 'Standart-Ansicht',
     'max_upcoming_days'     => 'Max. Tage für Bevorstehend',
     'use_locator'           => 'Locator-Plugin integrieren',
@@ -532,10 +533,10 @@ $LANG_configSelect['evlist'] = array(
         3 => 'Lesen-Schreiben',
     ),
     13 => array(
-        0 => 'Linke Blöcke',
-        1 => 'Rechte Blöcke',
-        2 => 'Linke & Rechte Blöcke',
-        3 => 'Keine',
+        0 => 'Keine',
+        1 => 'Linke Blöcke',
+        2 => 'Rechte Blöcke',
+        3 => 'Linke & Rechte Blöcke',
     ),
     14 => array(
         'day' => 'Tages-Ansicht',
@@ -546,39 +547,3 @@ $LANG_configSelect['evlist'] = array(
     ),
 );
 
-// Legacy config selections, requires glFusion 2.0
-$LANG_configselects['evlist'] = array(
-    0 => array('Ja' => 1, 'Nein' => 0),
-    1 => array('Ja' => TRUE, 'Nein' => FALSE),
-    2 => array('Keine' => 0, 'Veranstaltung hinzufügen' => 1, 'Veranstaltungs-Liste' => 2),
-    3 => array('Sonntag' => 1, 'Montag' => 2),
-    4 => array(
-            'Don Nov 20, 2008'      => '%a %b %d, %Y',
-            'Don Nov 20'            => '%a %b %d',
-            'Donnerstag Nov 20, 2008' => '%A %b %d, %Y',
-            'Donnerstag Nov 20'       => '%A %b %d',
-            'Donnerstag November 20'  => '%A %B %d',
-            'November 20, 2008'     => '%B %d, %Y',
-            '11/20/08'              => '%m/%d/%y',
-            '11-20-08'              => '%m-%d-%y',
-            '2008 11 20'            => '%Y %m %d',
-            'Nov 20 2008'           => '%b %d %Y',
-            'Nov 20, 2008'          => '%b %d, %Y',
-            '20 Nov. 2008'            => '%d %b. %Y'
-    ),
-    5 => array('02:38 PM' => '%I:%M %p', '14:48' => '%H:%M',),
-    6 => array(
-            'sobald Uhrzeit-Beginn um ist (wenn vorhanden)' => 1,
-            'sobald Datum-Beginn um ist (i.R. der nächste Tag)' => 2,
-            'sobald Uhrzeit-Ende um ist (wenn vorhanden)' => 3,
-            'sobald Datum-Ende um ist' => 4,
-        ),
-    7 => array('Oben auf Seite'=>1,'Nach Hauptartikel'=>2,'Unten auf Seite'=>3,'Gesamte Seite'=>0),
-    8 => array('Vergangen'=>1,'Bevorstehend'=>2,'Diese Woche'=>3,'Diesen Monat'=>4),
-    9 => array('Deaktiviert' => 0, 'Tabelle' => 1, 'Artikel' => 2, 'Kalender' => 3),
-    12 => array('Kein Zugriff' => 0, 'Nur-Lesen' => 2, 'Lesen-Schreiben' => 3),
-    13 => array('Linke Blöcke' => 0, 'Rechte Blöcke' => 1, 'Linke & Rechte Blöcke' => 2, 'Keine' => 3),
-    14 => array('Tages-Ansicht' => 'day', 'Wochen-Ansicht' => 'week', 'Monats-Ansicht' => 'month', 'Jahres-Ansicht' => 'year', 'Listen-Ansicht' => 'list'),
-    16 => array('HTML' => 'html', 'JSON' => 'json'),
-    17 => array('Nein' => 0, 'Standart Nein' => 1, 'Standart Bezahlt' => 2, 'Standart Bezahlt & Unbezahlt' => 3),
-);